requirements.txt
torch
transformers
deepface
vosk
librosa
nltk
SpeechRecognition
soundfile
opencv-python
tqdm
dostoevsky
ffmpeg-python
